+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- [bug](https://github.com/FrankC01/pysui/issues/244) Adding `epoch` property to SuiString for JSON RPC
+
 ### Changed
 
 - [enhancement](https://github.com/FrankC01/pysui/issues/241) Validator paging